refactor: Member 관련 파일 이동
