+++ src/main/java/com/moabam/api/domain/entity/enums/Role.java
@@ -1,4 +1,4 @@
-package com.moabam.api.domain;
+package com.moabam.api.domain.entity.enums;
 
 public enum Role {
